+++ Inventory.Frontend/Services/Interfaces/IOrderService.cs
@@ -5,7 +5,7 @@ namespace Inventory.Frontend.Services.Interfaces;
 public interface IOrderService
 {
     Task<IEnumerable<OrderViewModel>> GetOrdersAsync();
-    Task<OrderViewModel> GetOrderByIdAsync(long orderId);
+    Task<OrderViewModel> GetOrderByIdAsync(int orderId);
 
     Task CreateOrderAsync(OrderViewModel order);
     // Additional methods: Update, Delete, etc.
